chore(lint): eliminate file_length violations (#272)
Split five over-threshold files into sibling files so the `file_length`
rule's baseline contribution drops to zero:

- `App/MoolahApp.swift` → + `MoolahDomainCommands.swift` (4 Commands structs)
- `Backends/CloudKit/Migration/MigrationCoordinator.swift`
  → + `MigrationProfileNaming.swift` (pure naming helper enum)
- `Features/Import/Views/RecentlyAddedView.swift` → +
  `RecentlyAddedDetailSheet.swift` (read-only detail + rule-from-search sheet)
  and `RecentlyAddedNeedsSetupPanel.swift` (needs-setup / failed-files panel
  with its two rows)
- `Features/Import/Views/RuleEditorView.swift` → +
  `RuleEditorConditionRow.swift` (condition row + kind picker) and
  `RuleEditorActionRow.swift` (action row + kind picker)
- `Features/Settings/SettingsView.swift` → +
  `MoolahProfileDetailView.swift` (3 backend-specific profile detail views
  + shared auth status + iOS share sheet)

Regenerated `.swiftlint-baseline.yml` under the standing permission to
regen when violation counts strictly decrease:

| Rule | Before | After | Δ |
| --- | ---: | ---: | ---: |
| `file_length` | 5 | 0 | -5 |
| **Total** | **1002** | **997** | **-5** |

All other rules keep their existing counts; no new rules appear.

Fixes #272
